Added initial mock-up of collapsing leaf nodes
Basic idea is for certain clusters of leaf-siblings in the org chart
to 'collapse' them into a more readible layout by vertically aligning
them via graphviz magic.

Untested, and will probably have horrible bugs without some method
to reverse the changes as it temporarily makes the underlying graph
not a tree (which is something we assume elsewhere).